🚚structure: error code 타입 types/error.ts로 이동
